Blockchain Case Study
About  HIVE Blockchain Technologies Ltd:
HIVE Blockchain Technologies Ltd is in the Canadian owned business which provide infrastructure solutions in the blockchain industry. It focuses on the mining and sale of digital currencies, BTC,ETH and ETH classic. Its geographical segments include Canada, Sweden, Iceland, Switzerland, and Norway, out of which the majority of the revenue comes from Canada.

HIVE Blockchain Technologies Ltd which was formerly Leeta Gold Corp, is a Canada-based company within blockchain sector. HIVE partners with Genesis Mining Ltd, a cryptocurrency mining hashrate provider, in order to operate blockchain infrastructure, namely cryptocurrency mining facility in Iceland. The facility produces mined cryptocurrency Ethereum around the clock and is monitored with Genesis HIVE, Genesis' proprietary software tool for large-scale mining, to automatically optimize chip temperatures and power consumption needed for coin production. Also, the Company partners with Fiore Group, a Canadian merchant banking firm, in order to accelerate development of the blockchain sector through traditional capital markets.
How HIVE Works: 
HIVE’s operations provide shareholders with exposure to the operating margins of digital currency mining and a portfolio of coins. That makes this a unique global company out of the gate.
HIVE’S Business model:
HIVE is a relatively new public company which uses high powered computing assets to mine crypto currencies like Ethereum and Bitcoin on the cloud. Our deployments provide shareholders with exposure to the operating margins of digital currency mining as well as a portfolio of crypto-coins.
HIVE is a crypto currency mining firm. HIVE validate on blockchain networks. Everyday HIVE earns new crypto coins which it can monetize and cash flow. Crypto currency is extremely competitive, The largest inputs to mining is gas costs and hardware. Hive has reliable and cheap source of electricity and hardware.
HIVE business is totally different from the company who works for cloud mining, 
HIVE is one of raising block chain company in Canada, there business growing very fast.

Why HIVE successful crypto mining company:
Crypto currency is extremely competitive, hence the need for efficiencies of scale. The largest inputs to mining are electricity costs and hardware. You need a reliable and cheap source of electricity for your hardware. 
Fast and reliable internet access, as well as a cool climate, help to lower costs for crypto mining.
 
HIVE’s assets:
HIVE owns state-of-the-art green energy-powered data centre facilities in Canada, Sweden and Iceland, in which HIVE produce newly minted digital currencies like Bitcoin and Ethereum continuously on the cloud.
 

Which crypto currencies does HIVE mine:
HIVE currently mine Ethereum, Ethereum Classic and Bitcoin.
Proof of Mining of HIVE:
The breakthrough innovation in Satoshi Nakamoto’s whitepaper introducing Bitcoin, which set the stage for all cryptocurrencies to follow, was a mechanism for unrelated parties to securely transact and reliably come to consensus on the state of a transaction ledger (the blockchain) without having to trust a central intermediary.
This mechanism is known as proof-of-work mining, a system where a decentralized network of miners commit “work” — using electricity to power computer hardware — in exchange for receiving rewards in the form of cryptocurrency.
The original communications protocol for the Internet — HTTP — is still in use today notwithstanding the introduction of more efficient protocols. Just like HTTP, the original vision of Satoshi — proof-of-work mining — is the backbone of the whole blockchain world. After a decade and massive growth in total crypto market capitalization, this consensus mechanism has proven to be stable and highly secure, and has withstood numerous hacking attempts.
For these reasons, we expect proof-of-work mining and data centre infrastructure like HIVE’s will always be an essential component of the blockchain and cryptocurrency ecosystem.
Proof-of-stake of HIVE:
Proof-of-stake is an alternative method of fulfilling the same function as proof-of-work mining — the validation of cryptocurrency transactions — by having parties deposit, or “stake”, amounts of cryptocurrency into a smart contract.
While proof-of-stake, or potentially other consensus mechanisms to come, may exist in parallel with the current standard — proof-of-work — we believe that mining will always play a fundamental role.
The current proposal for Ethereum’s shift to proof of stake from proof of work has a number of unknown variables, including uncertainty over timing, execution and ultimate adoption; and there is not yet a definitive plan that is established and approved.  
If Ethereum’s shift to proof of stake were to occur, HIVE’s GPU-based mining provides flexibility to identify and focus computing resources on existing and new cryptocurrencies including those using the proof-of-work standard as they become more profitable to mine, as well as for other high efficiency computing applications, such as artificial intelligence and graphic rendering as well as private blockchain computing, which are becoming more widespread as the blockchain technology grows and develops.


About HIVE’s hardware:
Mining equipment is anticipated to remain useful for multiple years. HIVE benefits from monitoring software that constantly tracks things like hash rates, temperatures, and power consumption. This allows us to easily detect problematic rigs, and prevent overheating or hash rate drops, keeping maintenance costs to a minimum.

 

 

HIVE has no control over information available at or through hyperlinked sites. HIVE makes no representation as to, and is not responsible for, the quality, content, nature or reliability of any hyperlinked site, or any information available on or through a